requirements.txt
numpy==1.23.4
matplotlib==3.6.2
colorama==0.4.6
openai==0.27.2
requests==2.28.2
beautifulsoup4==4.12.1
python-pptx==0.6.21
wikipedia==1.4.0
wikipedia_sections
pyautogui==0.9.53
opencv-python==4.7.0.72
sv_ttk
pyperclip